-krypteret kode -implementeret changelog i alle (relevante) endpoints -tilføjet (endnu ikke implementeret) validerings funktioner

+++ src/main/java/p3project/MainController.java
@@ -9,6 +9,7 @@ import java.nio.file.Paths;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
+import java.time.LocalDate;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
@@ -28,6 +29,9 @@ import org.springframework.web.util.WebUtils;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+
+import org.mindrot.jbcrypt.BCrypt;
+
 import p3project.classes.Changelog;
 import p3project.classes.Contract;
 import p3project.classes.Eventlog;
@@ -104,7 +108,6 @@ public class MainController {
         if(storedSponsor == null) return renderSponsorPageWithResponse("Error retrieving sponsor. Please try again", model);
 
         return handleUpdateRequest(sponsor, storedSponsor, request, model);
-
     }
 
     @PostMapping("/update/contract")
@@ -117,7 +120,7 @@ public class MainController {
         //If a file is sent parse the data
         if(!pdffile.isEmpty()){
             parseContract(contract, pdffile);
-        }else{
+        } else {
             //if no pdf sent, get the stored values.
             contract.setPdfData(storedContract.getPdfData());
             contract.setFileName(storedContract.getFileName());
@@ -168,9 +171,11 @@ public class MainController {
                 Object before = field.get(storedObject);
                 Object after = field.get(requestObject);
                 if (!Objects.equals(before, after)) {
-                    User user = getUserFromToken(request);
-                    Changelog log = new Changelog(user, requestObject, field, before, after);
-                    logRepository.save(log);
+                    if(fieldShouldBeLogged(field)) {
+                        User user = getUserFromToken(request);
+                        Changelog log = new Changelog(user, requestObject, field, before, after);
+                        logRepository.save(log);
+                    }
                     field.set(storedObject, after);
                     fieldsChanged++;
                 }
@@ -187,7 +192,7 @@ public class MainController {
         return fieldsChanged;
     }
 
-    private boolean fieldShouldBeEvaluated(Field field) { // jank
+    private boolean fieldShouldBeLogged(Field field) { // jank
         String fieldName = field.getName();
         switch(fieldName) {
             case "pdfData":
@@ -206,7 +211,9 @@ public class MainController {
         User user = getUserFromToken(request);
         Eventlog log = new Eventlog(user, sponsor, "CREATED");
         logRepository.save(log);
+
         sponsorRepository.save(sponsor);
+
         return "redirect:/sponsors";
     }
 
@@ -216,6 +223,10 @@ public class MainController {
     public String addContractForSponsor(@ModelAttribute Contract contract, @RequestParam MultipartFile pdffile, Model model, HttpServletRequest request) {
         if(!userHasValidToken(request)) return "redirect:/login";
         if(!userIsAdmin(request))       return "redirect:/homepage";
+
+        User user = getUserFromToken(request);
+        Eventlog log = new Eventlog(user, contract, "CREATED");
+        logRepository.save(log);
 
         try {
             parseContract(contract, pdffile);
@@ -232,20 +243,68 @@ public class MainController {
         if(!userHasValidToken(request)) return "redirect:/login";
         if(!userIsAdmin(request))       return "redirect:/homepage";
 
+        User user = getUserFromToken(request);
+        Eventlog log = new Eventlog(user, service, "CREATED");
+        logRepository.save(log);
+
         try {
             serviceRepository.save(service);
             return "redirect:/sponsors";
         } catch (IllegalArgumentException ex) {
             return renderSponsorPageWithResponse("FEJL!!!", model);
         }
-
     }
+
+
+    private boolean sponsorIsValid(Sponsor sponsor) {
+        if(sponsor.getName() == null || sponsor.getName().isEmpty()) return false;
+
+        if(sponsor.getPhoneNumber().matches("[\\+\\-0-9]*")) return false;
+
+        if(sponsor.getCvrNumber().length() != 8
+        || !sponsor.getCvrNumber().matches("[0-9]*")
+        ) return false;
+
+        return true;
+    }
+
+    private boolean contractIsValid(Contract contract) {
+        if(contract.getName() == null || contract.getName().isEmpty()) return false;
+
+        if(contract.getPaymentAsInt() < 0) return false;
+
+        LocalDate start = contract.getStartDate();
+        LocalDate end = contract.getEndDate();
+        if(start != null && end != null && start.isAfter(end)) return false;
+
+        return true;
+    }
+
+    private boolean serviceIsValid(Service service) {
+        if(service.getName() == null || service.getName().isEmpty()) return false;
+
+        LocalDate start = service.getStartDate();
+        LocalDate end = service.getEndDate();
+        if(start != null && end != null && start.isAfter(end)) return false;
+
+        if(service.getAmountOrDivision() < 0) return false;
+
+        return true;
+    }
+
 
     // Deletes a service by ID
     @PostMapping("/sponsors/deleteService")
-    public String deleteService(@RequestParam Long serviceId, HttpServletRequest request) {
+    public String deleteService(@RequestParam Long serviceId, HttpServletRequest request, Model model) {
         if(!userHasValidToken(request)) return "redirect:/login";
         if(!userIsAdmin(request))       return "redirect:/homepage";
+
+        Service service = serviceRepository.findById(serviceId).orElse(null);
+        if(service == null) return renderSponsorPageWithResponse("Invalid id", model);
+
+        User user = getUserFromToken(request);
+        Eventlog log = new Eventlog(user, service, "DELETED");
+        logRepository.save(log);
 
         serviceRepository.deleteById(serviceId);
         return "redirect:/sponsors";
@@ -255,9 +314,16 @@ public class MainController {
 
     // Deletes a sponsor and all contracts linked to that sponsor
     @PostMapping("/sponsors/delete")
-    public String deleteSponsor(@RequestParam Long sponsorId, HttpServletRequest request) {
+    public String deleteSponsor(@RequestParam Long sponsorId, HttpServletRequest request, Model model) {
         if(!userHasValidToken(request)) return "redirect:/login";
         if(!userIsAdmin(request))       return "redirect:/homepage";
+
+        Sponsor sponsor = sponsorRepository.findById(sponsorId).orElse(null);
+        if(sponsor == null) return renderSponsorPageWithResponse("Invalid id", model);
+
+        User user = getUserFromToken(request);
+        Eventlog log = new Eventlog(user, sponsor, "DELETED");
+        logRepository.save(log);
 
         sponsorRepository.deleteById(sponsorId);
         Iterable<Contract> contracts = contractRepository.findAll();
@@ -272,9 +338,16 @@ public class MainController {
 
     // Deletes a contract by ID
     @PostMapping("/sponsors/deleteContract")
-    public String deleteContract(@RequestParam Long contractId, HttpServletRequest request) {
+    public String deleteContract(@RequestParam Long contractId, HttpServletRequest request, Model model) {
         if(!userHasValidToken(request)) return "redirect:/login";
         if(!userIsAdmin(request))       return "redirect:/homepage";
+
+        Contract contract = contractRepository.findById(contractId).orElse(null);
+        if(contract == null) return renderSponsorPageWithResponse("Invalid id", model);
+
+        User user = getUserFromToken(request);
+        Eventlog log = new Eventlog(user, contract, "DELETED");
+        logRepository.save(log);
         
         contractRepository.deleteById(contractId);
         return "redirect:/sponsors";
@@ -344,22 +417,20 @@ public class MainController {
 
 
     @PostMapping("/login/confirm")
-    public String confirmLogin(@RequestParam String username, @RequestParam String hashedPassword, @RequestParam boolean rememberMe, Model model, HttpServletResponse response) {
-        // String hashedPassword = bcrypt.hash(password);
-        // if(bcrypt.compare(hashedpassword, user.getPassword());
+    public String confirmLogin(@RequestParam String username, @RequestParam String password, @RequestParam boolean rememberMe, Model model, HttpServletResponse response) {
         User user = userRepository.findByName(username);
-        if(hashedPassword.equals(user.getPassword())) {
+        if(BCrypt.checkpw(user.getPassword(), password)) {
             String id = user.getId().toString();
             Token token = Token.sign(id);
             String formattedToken = id + "." + token.getHash();
             String encodedToken = URLEncoder.encode(formattedToken, StandardCharsets.UTF_8);
             int time = rememberMe ? (60 * 60 * 24 * 365) : (60 * 60 * 24); // 1 år vs 1 dag
             ResponseCookie cookie = ResponseCookie.from("token", encodedToken)
-            .httpOnly(true)  // noget med javascript
-            .secure(false)    // HTTPS only
-            .path("/")       // Bruges til alle sider i domænet
-            .maxAge(time)
-            .build();
+                .httpOnly(true)  // noget med javascript
+                .secure(false)    // HTTPS only
+                .path("/")       // Bruges til alle sider i domænet
+                .maxAge(time)
+                .build();
             response.addHeader("Set-Cookie", cookie.toString());
             return "redirect:/homepage";
         } else {
@@ -408,7 +479,6 @@ public class MainController {
         return user;
     }
 
-
     
     @GetMapping("/homepage")
     public String showhomepage(Model model, HttpServletRequest request) {
@@ -425,30 +495,23 @@ public class MainController {
 
     // Add user with demo sponsor & contract
     @PostMapping("/users/add")
-    public String addUserFromWeb(@RequestParam String name, @RequestParam String password) {
-        // String hashedPassword = bcrypt.hash(password)
-        // user.setPassword(hashedPassword);
+    public String addUserFromWeb(@RequestParam String name, @RequestParam String password, boolean isAdmin, Model model) {
+        List<User> users = userRepository.findAll();
+        for(User user : users) {
+            if(name.equals(user.getName())) {
+                model.addAttribute("responseMessage", "Username already exists. Please choose another.");
+                return "AdminPanel";
+            }
+        }
+        String hashedPassword = BCrypt.hashpw(password, BCrypt.gensalt());
         User user = new User();
         user.setName(name);
-        user.setPassword(password);
+        user.setPassword(hashedPassword);
+        user.setIsAdmin(isAdmin);
         userRepository.save(user);
 
         return "redirect:/users";
     }
-
-
-    //    public String addUserFromWeb(@RequestParam String name, @RequestParam String password, Model model) {
-    //     List<User> users = userRepository.findAll();
-    //     for(User user : users) {
-    //         if(name.equals(user.getName())) {
-    //             model.addAttribute("errorMessage", "Username already exists");
-    //             return "redirect:/AdminPanel";
-    //         }
-    //     }
-    //     User user = new User();
-    //     user.setName(name);
-    //     user.setPassword(password);
-
 
 
     @GetMapping("/testuser")
